// daily-app-frontend/Daily/src/app/rateplans/models/rateplan.model.ts
export interface Rateplan {
  activationSourceFlag: any;
  bundle: string;
  buzConsFlag: boolean;
  ciGrouping: string;
  combined: string;
  consumerFlag: boolean;
  contractType: string;
  dataVoiceAdslFlag: string;
  deactivationFlag: boolean;
  deferredFlag: boolean;
  enterpriseFlag: boolean;
  forIvrCost: string;
  forIvrRev: string;
  hybrid: boolean;
  loyalFlag: boolean;
  postFlag: boolean;
  postPreFlag: string;
  preFlag: boolean;
  ratePlan: string;
  ratePlanCode: string;
  ratePlanGroupKey: number;
  ratePlanKey: number;
  ratePlanType: string;
  segment: string;
  showFlag: boolean;
}